\name{MetropolisHastings}
\alias{MetropolisHastings}
\docType{package}
\title{
\packageTitle{MetropolisHastings}
}
\description{
\packageDescription{MetropolisHastings}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MetropolisHastings}
\packageIndices{MetropolisHastings}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MetropolisHastings}

Maintainer: \packageMaintainer{MetropolisHastings}
}
\references{
~~ Maria L. Rizzo: Statistical Computing with R, Chapter 9.2 ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
