\name{MetropolisHastings}
\alias{MetropolisHastings}
\docType{package}
\title{
\packageTitle{MetropolisHastings}
}
\description{
\packageDescription{MetropolisHastings}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MetropolisHastings}
\packageIndices{MetropolisHastings}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MetropolisHastings}

Maintainer: \packageMaintainer{MetropolisHastings}
}
\references{
~~ Maria L. Rizzo: Statistical Computing with R, Chapter 9.2 ~~
}

\keyword{ package }
\seealso{
}
\examples{
~~ simple examples of the most important functions ~~
}
